requirements.txt
faiss-cpu
sentence-transformers
pdfplumber
requests
tqdm
streamlit
arxiv
